// src/shared/api/models/getApiSalesDetailsId200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiSalesDetailsId200ShortInvoiceId } from './getApiSalesDetailsId200ShortInvoiceId';
import type { GetApiSalesDetailsId200CustomerId } from './getApiSalesDetailsId200CustomerId';
import type { GetApiSalesDetailsId200CustomerName } from './getApiSalesDetailsId200CustomerName';
import type { GetApiSalesDetailsId200Tips } from './getApiSalesDetailsId200Tips';
import type { GetApiSalesDetailsId200SalesNote } from './getApiSalesDetailsId200SalesNote';
import type { GetApiSalesDetailsId200VoidId } from './getApiSalesDetailsId200VoidId';
import type { GetApiSalesDetailsId200SalesType } from './getApiSalesDetailsId200SalesType';
import type { GetApiSalesDetailsId200CreatedBy } from './getApiSalesDetailsId200CreatedBy';
import type { GetApiSalesDetailsId200UpdatedBy } from './getApiSalesDetailsId200UpdatedBy';
import type { GetApiSalesDetailsId200SellBy } from './getApiSalesDetailsId200SellBy';
import type { GetApiSalesDetailsId200SyncedWith } from './getApiSalesDetailsId200SyncedWith';
import type { GetApiSalesDetailsId200SyncId } from './getApiSalesDetailsId200SyncId';
import type { GetApiSalesDetailsId200ItemsItem } from './getApiSalesDetailsId200ItemsItem';
import type { GetApiSalesDetailsId200TransactionsItem } from './getApiSalesDetailsId200TransactionsItem';
import type { GetApiSalesDetailsId200TaxItem } from './getApiSalesDetailsId200TaxItem';
import type { GetApiSalesDetailsId200SalesDiscountsItem } from './getApiSalesDetailsId200SalesDiscountsItem';
import type { GetApiSalesDetailsId200Customer } from './getApiSalesDetailsId200Customer';
import type { GetApiSalesDetailsId200SalesSync } from './getApiSalesDetailsId200SalesSync';

export type GetApiSalesDetailsId200 = {
    id: string;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    seqId: number;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    invAutoNumber: number;
    salesDate: string;
    invoiceId: string;
    shortInvoiceId: GetApiSalesDetailsId200ShortInvoiceId;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    outletId: number;
    customerId: GetApiSalesDetailsId200CustomerId;
    customerName: GetApiSalesDetailsId200CustomerName;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    subTotal: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    totalTax: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    totalDiscount: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    netTotal: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    tenderAmount: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    change: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    roundOff: number;
    tips: GetApiSalesDetailsId200Tips;
    salesNote: GetApiSalesDetailsId200SalesNote;
    voidId: GetApiSalesDetailsId200VoidId;
    salesType: GetApiSalesDetailsId200SalesType;
    createdBy: GetApiSalesDetailsId200CreatedBy;
    updatedBy: GetApiSalesDetailsId200UpdatedBy;
    sellBy: GetApiSalesDetailsId200SellBy;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    sourceId: number;
    syncedWith: GetApiSalesDetailsId200SyncedWith;
    syncId: GetApiSalesDetailsId200SyncId;
    createdAt: string;
    updatedAt: string;
    items: GetApiSalesDetailsId200ItemsItem[];
    transactions: GetApiSalesDetailsId200TransactionsItem[];
    tax: GetApiSalesDetailsId200TaxItem[];
    salesDiscounts: GetApiSalesDetailsId200SalesDiscountsItem[];
    customer?: GetApiSalesDetailsId200Customer;
    paidAmount?: number;
    salesSync?: GetApiSalesDetailsId200SalesSync;
};